Version 1.91
Map files now specify the world name they're associated with to allow
different worlds to use different maps with the same EmpireHUD folder
Added command: gui mapurl <main/alt> <url>: This command temporarily
changes the URL to automatically download MUD maps from.
Reformatted the map's right click menu a bit
The map now shows an arrow to indicate the direction to the player if it
is panned so that the player is no longer visible.
Fixed some unhandled errors if a map is not found.